face-recognition-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
Pillow==10.1.0
numpy==1.25.2
python-jose==3.3.0
passlib==1.7.4
python-bcrypt==0.4.4
